Refactor DamageSystem, burst fire for Explosives
Damage cooldown is now per entity and not per entityType, which makes
explosives a better weapon and for it to be a better companion of
rockets.

Had to add mutex for every entity, might be a big load on memory,
I will have an eye on this.

Add damage for asteroid-player collision. Need to work on GameOverScreen
and GameOver state.