Add coins and GameSessionManager to track them during game session
Did a lot of things here. I added the GameSessionManager that should
handle numbers for the current game session, such as gold and
asteroids destroyed, for now.

Did a small refactor on includes for Systems in game.hpp, game.cpp
and screens.

Add gold total in MainMenuScreen.
Add gold and enemies in PlayingScreen.