Add UpgradeStoreScreen and Button component
Refactor a lot of how clicks work.

Refactor GUISystem to ScreenManager, which fits better.

Add GUIInteractionSystem that is in charge of updating gui state
components.

Add GUIStateComponent, ButtonComponent, SliderComponent, LabelComponent.
This components will serve as base for all GUI entities.

Right now the GUIState is behaving weirdly, but didn't want to
take longer on this commit. I'll fix that as a bug.

GUIComponent now holds the position for the GUI entities. No
Transform component is needed for GUI entities.

Will work on updating the RenderSystem so it takes into consideration
the parent position and the position of the child GUI entity.