Add TextureManager
Modify the size of the ship to always be 50x50 and do not depend on
texture size.

Start implementing the TextureManager, I'll work on a PackageReader
next, so we can now obfuscate textures for release version.

Need to modify how ALL textures are loaded, as now they are accesible
everywhere with the TextureManager