Add PackReader for managing assets from .pak
Still need to work on sounds and fonts, but they should be easily
modified to be read from .pak

Modify renderEx to now do colorMod there and not modify all
entities that use that texture. This is working weirdly on WSL,
I'll make sure is working correctly on Windows. Similar issue with
audio device, I'm getting leaks and errors on WSL but everything is
working fine in native Windows.

Modify how textures are loaded on several places (I think it's all
of them but I might have missed something. Textures for texts are
still loaded for each screen that needs that text).