Add Android Compatibility
You can now open the project on AndroidStudio, just open the `android`
folder.
It should open and show the MainScreen. Touch input is working as well.
Still need to work on JoystickTouch or something like that.

Screen size/resolution still not working correctly on Android.

Committing right now because this commit was getting really big already.