Refactor Animations!
Animations will ahve to share name i.e. shipDamage001, shipDamage002, etc
Then can be loaded with TextureManager::instance().getAnimationFrames(name)
This will return a vector of Textures to be placed in the animation component
of the entity.

I'm thinking on adding possibility of having different time for each
frame in an animation.